// web/src/api/generated/types/ComponentResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ComponentProperties } from "./ComponentProperties.ts";

/**
 * ComponentResponse
 * @description Model for containing the response of getting the component dictionary.
 * @example [object Object]
*/
export type ComponentResponse = {
    /**
     * @description Dictionary of component_ids as string and ComponentProperties as value
     * @type object
    */
    components: {
        [key: string]: ComponentProperties;
    };
};