// web/src/api/generated/types/MultiflashResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


/**
 * MultiflashResponse
 * @example [object Object]
*/
export type MultiflashResponse = {
    /**
     * @description Phase labels (vapor, liquid, aqueous, mercury) with their fraction of unity and mercury concentration
     * @type object
    */
    phaseValues: {
        [key: string]: {
            [key: string]: number;
        };
    };
    /**
     * @description Mole fractions of each of the components (Note: mass is discarded from MultiflashResult)
     * @type object
    */
    componentFractions: {
        [key: string]: number[];
    };
    /**
     * @description Ratio of components in the feed (guaranteed to sum to 1)
     * @type object
    */
    feedFractions: {
        [key: string]: number;
    };
};